# backend/control_plane/openapi.yaml
openapi: 3.0.0
info:
  title: Remind Me
  description: API для управления напоминаниями, привычками и интеграциями с Telegram и календарем.
  version: 0.0.1
servers:
  - url: https://localhost:9000/v1
    description: Основной сервер
paths:
...
  /user:
    get:
      summary: Получение информации о текущем пользователе
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      security:
        - BearerAuth: []
    put:
      summary: Обновление информации о текущем пользователе
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      security:
        - BearerAuth: []
  /reminders:
    post:
      summary: Создание нового напоминания
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reminder'
      responses:
        '201':
          description: Напоминание успешно создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
      security:
        - BearerAuth: []

    get:
      summary: Получение списка напоминаний
      parameters:
        - name: tag
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, forgotten]
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Список напоминаний
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reminder'
      security:
        - BearerAuth: []
  /reminders/{reminderId}:
    put:
      summary: Редактирование напоминания
      parameters:
        - name: reminderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reminder'
      responses:
        '200':
          description: Напоминание успешно обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
      security:
        - BearerAuth: []

    delete:
      summary: Удаление напоминания
      parameters:
        - name: reminderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Напоминание успешно удалено
      security:
        - BearerAuth: []
  /reminders/{reminderId}/complete:
    post:
      summary: Отметка напоминания как выполненного
      parameters:
        - name: reminderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Напоминание отмечено как выполненное
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
      security:
        - BearerAuth: []
  /reminders/{reminderId}/postpone:
    post:
      summary: Откладывание напоминания
      parameters:
        - name: reminderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newTime:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Напоминание отложено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
      security:
        - BearerAuth: []
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        sex:
          type: string
          enum: [male, female]
        firstName:
          type: string
        secondName:
          type: string
        birthDate:
          type: string
          format: date
        telegramId:
          type: string
        calendarIntegrationKey:
          type: string
        timezone:
          type: string
        level:
          type: integer
        experience:
          type: integer
        streak:
          type: integer
        createdAt:
          type: string
          format: date-time
        lastActive:
          type: string
          format: date-time
    Reminder:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        text:
          type: string
        time:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, completed, forgotten]
        removed:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        notificationSent:
          type: boolean
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT